requirements.txt
Flask==3.1.1
beautifulsoup4==4.13.4
requests==2.32.5
google-cloud-secret-manager==2.24.0